Refactor: "NotFound Exception 커스텀 예외처리 설계"
자주 쓰는 NOTFOUND 예외 던지는 상황에서 구체적인 로깅을 위한 커스텀 예외 설계

Resolves: #39
