+++ src/main/java/umc/liview/exception/NotFoundException.java
@@ -1,0 +1,12 @@
+package umc.liview.exception;
+
+import umc.liview.exception.code.ErrorCode;
+
+public class NotFoundException extends BusinessException{
+    public NotFoundException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+    public NotFoundException(ErrorCode errorCode, long id) {
+        super(errorCode, "id " + id + " is not found");
+    }
+}
